// resources/js/types/candidate.ts
export interface Candidate {
    id_candidato_anio: number;
    id_candidato: number;
    id_anio: number;
    id_estado_aprobacion: number;
    fecha_aprobacion?: string | Date;
    destacado: number;
    motivo_rechazo?: string;
    id_tipo_eleccion: number;
    id_departamento?: number;
    id_provincia?: number;
    id_distrito?: number;
    id_centro_poblado?: number;
    id_cargo: number;
    numero_lista?: number;
    condicion_candidato: 'afiliado' | 'designado';
    propuestas?: string;
    foto_url?: string;
    cv_url?: string;
    creado_por?: number;
    creado_en: string | Date;
    actualizado_por?: number;
    actualizado_en?: string | Date;
    eliminado: number;
    eliminado_por?: number;
    eliminado_en?: string | Date;
    id_ubicacion_virtual?: string;
    nombre_completo?: string;
    nombre_cargo?: string;
    nombre_departamento?: string;
    ocupacion?: string;
    correo?: string;
    celular?: string;
    telefono?: string;
    nombre_provincia?: string;
    nombre_distrito?: string;
    nombre_centro_poblado?: string;
    propuestas_detalladas?: { titulo: string; descripcion_completa: string; imagen_referencia_url?: string }[];
    redes_sociales?: { nombre: string; icono_class: string; url: string }[];
}